+++ master/scenecompiler.py
@@ -233,7 +233,7 @@ class Brush:
 
     def __str__(self):
         return """
-        float sd{}( double3 v){{
+        float sd{}( float3 v){{
 
             {}
 
@@ -250,7 +250,7 @@ class Material:
 
     def __str__(self):
         return """
-        double3 shader{} (double3 gv, double3 lv, double3 n){{
+        float3 shader{} (float3 gv, float3 lv, float3 n){{
 
             {}
 
@@ -426,9 +426,9 @@ class _SceneCompiler:
         self.abs_normals=self.define_material(body = "return fabs(n);")
         self.basic_lighting = self.define_material(body = """
         
-        double3 n_g = n.x*rgt_g+n.y*upp_g+n.z*fwd_g;
+        float3 n_g = n.x*rgt_g+n.y*upp_g+n.z*fwd_g;
 
-        float L = dot(n_g,(double3)(0.0,0.0,-1.0)); return (double3)(L,L,L);
+        float L = dot(n_g,(float3)(0.0,0.0,-1.0)); return (float3)(L,L,L);
 
 
 
@@ -486,7 +486,7 @@ class _SceneCompiler:
         {}
 
 
-        float sdf_bank(double3 v, unsigned char shape_id){{
+        float sdf_bank(float3 v, unsigned char shape_id){{
 
             switch(shape_id){{
 
@@ -498,7 +498,7 @@ class _SceneCompiler:
 
         }}
 
-        double3 shader_bank(double3 gv, double3 lv, double3 n, unsigned char material_id){{
+        float3 shader_bank(float3 gv, float3 lv, float3 n, unsigned char material_id){{
 
 
             switch(material_id){{
@@ -507,7 +507,7 @@ class _SceneCompiler:
 
             }}
 
-            return (double3)(1.0, 1.0, 1.0);
+            return (float3)(1.0, 1.0, 1.0);
         }}
